+++ CHANGES.md
@@ -18,6 +18,7 @@ Changes in 4.8
 - JS optimize jqplotToColorBox() function in jquery.jqplottocolorbox.js
 - Use TinyMCE input for Studies Certificate text in Transcripts.php
 - Move Studies Certificate title and Signatures HTML to Template in Transcripts.php, rosariosis.sql, rosariosis_es.sql & rosariosis_fr.sql
+- Add TranscriptsIncludeForm() function in Transcripts.fnc.php & Transcripts.php
 
 Changes in 4.7.1
 ----------------